pyproject.toml
[project]
name = "spotify-to-deezer"
version = "0.1.0"
description = "Migration des playlists et favoris Deezer vers Spotify"
authors = [{name = "AutoGenerated"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deezer-python>=5.2.1",
    "spotipy>=2.23.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = ["black>=23.12.1", "mypy>=1.8.0"]

[project.scripts]
spotify-to-deezer = "spotify_to_deezer.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
